Added capacity for parsing ad topics
